% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{plot.bands}
\alias{plot.bands}
\title{Plots prediction bands for new observations}
\usage{
\method{plot}{bands}(bands, ynew = NULL)
}
\arguments{
\item{bands}{# Predicted bands fitted using \code{\link{predict.predictionBands}}}

\item{ynew}{# (optional) response values of new samples}
}
\value{
# ggplot object with a plot of the predicition bands (together with ynew if these are available)
}
\description{
Plots prediction bands for new observations
}
\examples{
 See \code{\link{fit_predictionBands}}
}
