% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{predict.predictionBands}
\alias{predict.predictionBands}
\title{Compute conformal prediction bands based on density estimation on new samples}
\usage{
\method{predict}{predictionBands}(cd_split_fit, xnew, type = "dist", alpha = 0.1)
}
\arguments{
\item{cd_split_fit}{# Object of the class predictionBands, fitted using function \code{\link{fit_predictionBands}}}

\item{xnew}{# new covariates (one per row) where prediction bands are to be computed}

\item{type}{# type of prediction bands. dist for 'dist-split' and cd for 'cd-split'.}

\item{alpha}{# Miscoverage level (defaults to 10\%)}
}
\value{
Object of the class 'bands' with the following components:
\item{y_grid}{Grid of values for y}
\item{densities}{Matrix where each row contains the estimated density of the corresponding row of xnew on each value of y_grid}
\item{ths}{thresholds agains which to compare conformal scores.}
\item{prediction_bands_which_belong}{List with logical values with the same size as nrow(xnew); each element contains shows which elements of y_grid are contained on the prediction bands of the corresponding row of xnew}
\item{intervals}{List with the same size as nrow(xnew) that contains the prediction bands for each element of xnew. For cd-split, results are shown as a union of disjoint intervals.}
\item{type}{Type of prediction band}
\item{alpha}{Miscoverage level}
}
\description{
Compute conformal prediction bands based on density estimation on new samples
}
\examples{
See \code{\link{fit_predictionBands}}
}
